[refactor] move a implementation that collects GPS location to MainViewModel class.
