+++ app/src/main/java/ymse3p/app/audiorecorder/util/MyException.kt
@@ -4,5 +4,9 @@ import java.lang.RuntimeException
 
 class CannotSaveAudioException(message: String) : RuntimeException(message) {
 }
+
 class CannotStartRecordingException(message: String) : RuntimeException(message) {
+}
+
+class CannotCollectGpsLocationException(message: String) : RuntimeException(message) {
 }
